InventoryCreature class

+++ app/src/main/java/com/dragons/carmenddragonsinventory/InventoryCreature.java
@@ -40,11 +40,11 @@ public class InventoryCreature {
         return imgFilePath;
     }
 
-    public float getCostToProduce() {
+    public double getCostToProduce() {
         return costToProduce;
     }
 
-    public float getListPrice() {
+    public double getListPrice() {
         return listPrice;
     }
